Restructure directories and remove unneeded files.
