+++ src/App.js
@@ -1,6 +1,6 @@
 import logo from './assets/map-icon.svg';
-import SearchForm from './SearchForm.js';
-import LocationDataDisplay from './LocationDataDisplay';
+import SearchForm from './components/SearchForm.js';
+import LocationDataDisplay from './components/LocationDataDisplay';
 import Image from 'react-bootstrap/Image';
 import React from 'react';
 import axios from 'axios';
